requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
influxdb-client==1.39.0
pymongo==4.6.1
neo4j==5.15.0
pycryptodome==3.19.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
django-prometheus==2.3.1
django-health-check==3.17.0
django-storages==1.14.2
django-guardian==2.4.0
django-multiselectfield==0.1.12
django-graphql-jwt==0.4.0
graphene-django==3.1.5 